requirements.txt
numpy
keras
tensorflow-gpu
sklearn
pandas
librosa
autopool